prediction of author of program
Code uses   62 programmers .arff   that  has  62
classes  and  7879features.The  feature vectors  are  sparse.  Each one
of  the  classes  corresponds  to  a  C++ programmer and  has  9
instances.  Each  instance  is  a  numeric  representation  of  one
piece  of source  code.  The  machine  learning  task  is  correctly
identifying  the  programer of anonymous  source  code